% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{boundaryAlphaHull}
\alias{boundaryAlphaHull}
\title{Create Alpha hull to identify grooves}
\usage{
boundaryAlphaHull(
  boundaryW,
  outRegionBoundary,
  alpha = 10,
  pointsCol = "red",
  hullCol = "blue"
)
}
\arguments{
\item{boundaryW}{boundary_cimg in white mode}

\item{outRegionBoundary}{one set of the boundary points from outRegion function}

\item{alpha}{alpha used for generating alpha hull}

\item{pointsCol}{color to display boundary points}

\item{hullCol}{color to display the hull}
}
\description{
Create Alpha hull to identify grooves
}
