% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{outRegion}
\alias{outRegion}
\title{Find highlighted regions}
\usage{
outRegion(boundary_cimg, maxY = 1.5, px = NA)
}
\arguments{
\item{maxY}{maximum???}

\item{px}{number of pixels}

\item{cimg}{A cimg from boundary_cimg output}
}
\description{
A helper function to determine the number of pixels in highlight regions
}
